@@ -2,14 +2,16 @@
 
 <PageTitle>Animal Matching Game</PageTitle>
 
-<style> 
+<style>
+
     .container {
-        width: 400px; 
+        width: 400px;
     }
-        button {
+
+    button {
         width: 100px;
         height: 100px;
-        font-size: 50px; 
+        font-size: 50px;
     }
 </style>
 
@@ -19,7 +21,7 @@
         {
             <div class="col-3">
                 <button type="button"
-                  class="btn btn-outline-dark">
+                        class="btn btn-outline-dark">
                     <h1>@animal</h1>
                 </button>
             </div>
@@ -28,15 +30,15 @@
 </div>
 
 @code {
-    List<string> animalEmoji = new List<string>()
-    {
-        "🐶","🐶",
-        "🐺","🐺",
-        "🐮","🐮",
-        "🦊","🦊",
-        "🐱","🐱",
-        "🦁","🦁",
-        "🐯","🐯",
-        "🐹","🐹",
-    };
+    List<string> animalEmoji = new()
+  {
+    "🐶","🐶",
+    "🐺","🐺",
+    "🐮","🐮",
+    "🦊","🦊",
+    "🐱","🐱",
+    "🦁","🦁",
+    "🐯","🐯",
+    "🐹","🐹",
+  };
 }
