@@ -2,31 +2,33 @@
 
 <PageTitle>Animal Matching Game</PageTitle>
 
-<style> 
+<style>
+
     .container {
-        width: 400px; 
+        width: 400px;
     }
-        button {
+
+    button {
         width: 100px;
         height: 100px;
-        font-size: 50px; 
+        font-size: 50px;
     }
 </style>
 
 <div class="container">
     <div class="row">
-        @for (var animalNumber = 0; 
-              animalNumber < shuffledAnimals.Count; 
-              animalNumber++)
+        @for (var animalNumber = 0;
+        animalNumber < shuffledAnimals.Count;
+        animalNumber++)
         {
             var animal = shuffledAnimals[animalNumber];
-            var uniqueDescription = 
-                   $"Button #{animalNumber}";
+            var uniqueDescription =
+            $"Button #{animalNumber}";
             <div class="col-3">
-                <button type="button" 
-                @onclick="@(() => 
-                    ButtonClick(animal, uniqueDescription))"
-                class="btn btn-outline-dark">
+                <button type="button"  
+                @onclick="@(() => 
+                    ButtonClick(animal, uniqueDescription))"
+                                class="btn btn-outline-dark">
                     <h1>@animal</h1>
                 </button>
             </div>
@@ -35,20 +37,9 @@
     <div class="row">
         <h2>Matches found: @matchesFound</h2>
     </div>
-    <div class="row">
-        <h2>Time: @timeDisplay</h2>
-    </div>
 </div>
 
-
-@using System.Linq
-@using System.Timers
-
 @code {
-
-    Timer timer = new Timer(100);
-    int tenthsOfSecondsElapsed = 0;
-    string timeDisplay = "";
 
     List<string> shuffledAnimals = new();
     int matchesFound = 0;
@@ -56,12 +47,9 @@
     private void SetUpGame()
     {
         shuffledAnimals = animalEmoji
-            .OrderBy(item => Random.Shared.Next())
-            .ToList();
+          .OrderBy(item => Random.Shared.Next())
+          .ToList();
         matchesFound = 0;
-        timer = new Timer(100);
-        timer.Elapsed += Timer_Elapsed;
-        tenthsOfSecondsElapsed = 0;
     }
 
     protected override void OnInitialized()
@@ -69,17 +57,17 @@
         SetUpGame();
     }
 
-    List<string> animalEmoji = new List<string>()
-    {
-        "🐶","🐶",
-        "🐺","🐺",
-        "🐮","🐮",
-        "🦊","🦊",
-        "🐱","🐱",
-        "🦁","🦁",
-        "🐯","🐯",
-        "🐹","🐹",
-    };
+    List<string> animalEmoji = new()
+  {
+    "🐶","🐶",
+    "🐺","🐺",
+    "🐮","🐮",
+    "🦊","🦊",
+    "🐱","🐱",
+    "🦁","🦁",
+    "🐯","🐯",
+    "🐹","🐹",
+  };
 
     string lastAnimalFound = string.Empty;
     string lastDescription = string.Empty;
@@ -90,49 +78,27 @@
             // First selection of the pair. Remember it.
             lastAnimalFound = animal;
             lastDescription = animalDescription;
-            timer.Start();
         }
-
         else if ((lastAnimalFound == animal)
-                 && (animalDescription != lastDescription))
+              && (animalDescription != lastDescription))
         {
             // Match found! Reset for next pair.
             lastAnimalFound = string.Empty;
 
             // Replace found animals with empty string to hide them.
             shuffledAnimals = shuffledAnimals
-                .Select(a => a.Replace(animal, string.Empty))
-                .ToList();
-
+          .Select(a => a.Replace(animal, string.Empty))
+          .ToList();
             matchesFound++;
             if (matchesFound == 8)
             {
-                timer.Stop();
-                timeDisplay += " - Play again?";
                 SetUpGame();
             }
         }
-
         else
         {
             // User selected a pair that don't match, reset selection.
             lastAnimalFound = string.Empty;
         }
     }
-
-    private void Timer_Elapsed(object? source,
-                        ElapsedEventArgs e)
-    {
-        InvokeAsync(
-            () =>
-            {
-                tenthsOfSecondsElapsed++;
-                timeDisplay =
-                       (tenthsOfSecondsElapsed / 10F)
-                           .ToString("0.0s");
-                StateHasChanged();
-            }
-        );
-    }
-
 }
