@@ -12,7 +12,7 @@
         </div>
         <div class="col-2">
             <input type="text" class="form-control"
-            @bind="q.Answer"/>
+            @bind="q.Answer" />
         </div>
         <div class="col-6">
             @if (q.Check())
@@ -22,6 +22,7 @@
         </div>
     </div>
 </div>
+
 
 @code {
     private Question q = new();
