@@ -6,37 +6,45 @@
     <div class="row">
         <h1>Math quiz!</h1>
     </div>
-    @foreach (Question q in questions) {
+    @foreach (Question q in questions)
+    {
         <div class="row mt-2">
             <div class="col-2">
                 @q.N1 @q.Operator @q.N2 =
             </div>
             <div class="col-2">
-                <input type="text" class="form-control"
-                @bind="q.Answer"/>
+                <input type="text"
+                       class="form-control" @bind="q.Answer" />
             </div>
             <div class="col-6">
                 @if (q.Check())
                 {
-                    <span>Correct!</span>
+                    <p>Correct!</p>
                 }
             </div>
         </div>
     }
+
+    @code {
+        private int NumberCorrect() =>
+        questions.Where(q => q.Check()).Count();
+    }
+
     <div class="row mt-5">
         <div class="col-6">
             @{
                 var numberCorrect = NumberCorrect();
             }
-            @if(numberCorrect < QUESTION_COUNT)
+            @if (numberCorrect < QUESTION_COUNT)
             {
                 <h6>
-                    @numberCorrect out of @QUESTION_COUNT
+                    @numberCorrect correct out of @QUESTION_COUNT
                 </h6>
-            } else
+            }
+            else
             {
                 <button class="btn btn-primary"
-                        @onclick="GenerateQuestions">
+                @onclick="GenerateQuestions">
                     Generate New Quiz
                 </button>
             }
@@ -44,12 +52,8 @@
     </div>
 </div>
 
-@using System.Linq
 
 @code {
-    private int NumberCorrect() =>
-            questions.Where(q => q.Check()).Count();
-
     private List<Question> questions = new();
     private const int QUESTION_COUNT = 6;
 
