@@ -10,34 +10,25 @@
     </div>
     <div class="row mt-2">
         <div class="col">
-            Enter text:
+            Enter text: 
         </div>
         <div class="col">
             <input type="text" class="form-control"
-            @onchange="UpdateValue"/>
+            @onchange="UpdateValue" />
         </div>
     </div>
     <div class="row mt-5">
-        <div class="col">
-            <h2>
-                Here's the value: <strong>@displayValue</strong>
-            </h2>
-        </div>
+        <h2>
+            Here's the value: <strong>@displayValue</strong>
+        </h2>
     </div>
     <div class="row mt-2">
         <div class="col">
-            Pick a number:
+            Pick a number: 
         </div>
         <div class="col">
-            <input type="range" class="form-control" 
-            @onchange="UpdateNumericValue"/>
-        </div>
-    </div>
-    <div class="row mt-5">
-        <div class="col">
-            <h2>
-                Here's the value: <strong>@displayValue</strong>
-            </h2>
+            <input type="range" class="form-control"
+            @onchange="UpdateNumericValue" />
         </div>
     </div>
 </div>
@@ -58,5 +49,4 @@
             displayValue = result.ToString();
         }
     }
-
 }
